split ps1 file
