+++ src/dra/692E8.c
@@ -301,10 +301,11 @@ void func_80109990(void) {
     }
 }
 
-static void CheckFloor(void);
-static void CheckCeiling(void);
-static void CheckWallRight(void);
-static void CheckWallLeft(void);
+void CheckFloor(void);
+void CheckCeiling(void);
+void CheckWallRight(void);
+void CheckWallLeft(void);
+
 static void CheckStageCollision(s32 isTransformed) {
     s32 i;
     s32 psp_s2;
@@ -1319,629 +1320,5 @@ block_160:
         FntPrint("step:%04x\n", PLAYER.step);
         FntPrint("bat_i_step:%04x\n", g_Player.unk66);
 #endif
-    }
-}
-
-void GetPlayerSensor(Collider* col) {
-    void (*theFunc)(Collider* col);
-    u32 mod = 0;
-    if (g_PlayableCharacter == PLAYER_ALUCARD) {
-        if(g_Player.status & PLAYER_STATUS_WOLF_FORM){
-            mod += 2;
-        }
-        col->unk14 = g_SensorsWall[0].x - mod;
-        col->unk1C = g_SensorsWall[0].y + mod;
-        col->unk18 = g_SensorsFloor[1].y - 1;
-        col->unk20 = g_SensorsCeiling[1].y + 1;
-    } else {
-        theFunc = g_PlOvl.GetPlayerSensor;
-        theFunc(col);
-    }
-}
-
-void func_8010BFFC(void) {
-    s16 x;
-    s16 y;
-    s32 i;
-    bool collided = false;
-    u16 original_Y = PLAYER.posY.i.hi;
-
-    if ((g_Player.pl_vram_flag & 1) || (D_801396EA != 0) ||
-#if defined(VERSION_US)
-        (g_Player.status &
-         (PLAYER_STATUS_TRANSFORM | PLAYER_STATUS_UNK40000000)) ||
-        ((g_PlayableCharacter != 0) && (PLAYER.step == 0x18)) ||
-#elif defined(VERSION_HD)
-        (g_Player.status & PLAYER_STATUS_TRANSFORM) ||
-#endif
-        (g_Player.unk78 == 1)) {
-        return;
-    }
-    if (PLAYER.posY.i.hi < 48) {
-        PLAYER.posY.i.hi -= 16;
-        while (true) {
-            for (i = 0; i < 4; i++) {
-                x = PLAYER.posX.i.hi + g_SensorsFloor[i].x;
-                y = PLAYER.posY.i.hi + g_SensorsFloor[i].y;
-                CheckCollision(x, y, &g_Player.colFloor[i], 0);
-            }
-            if ((g_Player.colFloor[1].effects &
-                 (EFFECT_SOLID_FROM_BELOW + EFFECT_SOLID)) == EFFECT_SOLID ||
-                (g_Player.colFloor[2].effects &
-                 (EFFECT_SOLID_FROM_BELOW + EFFECT_SOLID)) == EFFECT_SOLID ||
-                (g_Player.colFloor[3].effects &
-                 (EFFECT_SOLID_FROM_BELOW + EFFECT_SOLID)) == EFFECT_SOLID) {
-                // I don't know man
-                (*(&PLAYER)).posY.i.hi--;
-                PLAYER.velocityY = 0;
-                collided = true;
-                continue;
-            }
-            if (collided) {
-                return;
-            }
-            PLAYER.posY.i.hi += 8;
-            if (PLAYER.posY.i.hi >= 48) {
-                PLAYER.posY.i.hi = original_Y;
-                return;
-            }
-        }
-    } else if (PLAYER.posY.i.hi > 176) {
-        PLAYER.posY.i.hi += 32;
-        while (true) {
-            for (i = 0; i < 4; i++) {
-                x = PLAYER.posX.i.hi + g_SensorsCeiling[i].x;
-                y = PLAYER.posY.i.hi + g_SensorsCeiling[i].y;
-#if defined(VERSION_US)
-                if (g_PlayableCharacter != PLAYER_ALUCARD) {
-                    y += 6;
-                }
-#endif
-                CheckCollision(x, y, &g_Player.colFloor[4 + i], 0);
-            }
-            if ((g_Player.colCeiling[1].effects &
-                 (EFFECT_SOLID_FROM_ABOVE + EFFECT_SOLID)) == EFFECT_SOLID ||
-                (g_Player.colCeiling[2].effects &
-                 (EFFECT_SOLID_FROM_ABOVE + EFFECT_SOLID)) == EFFECT_SOLID ||
-                (g_Player.colCeiling[3].effects &
-                 (EFFECT_SOLID_FROM_ABOVE + EFFECT_SOLID)) == EFFECT_SOLID) {
-                // I don't know man
-                (*(&PLAYER)).posY.i.hi++;
-                PLAYER.velocityY = 0;
-                collided = true;
-                continue;
-            }
-            if (collided) {
-                PLAYER.posY.i.hi--;
-                return;
-            }
-            PLAYER.posY.i.hi -= 8;
-            if (PLAYER.posY.i.hi <= 176) {
-                PLAYER.posY.i.hi = original_Y;
-                return;
-            }
-        }
-    }
-}
-
-static void CheckFloor(void) {
-    Collider sp10;
-    byte stackpad[40];
-    s32 var_s1;
-    s32 var_a0;
-    s32 var_s2;
-    s32 sp30;
-    s32 var_s8;
-    s16 var_s6;
-    s16 argX;
-    s16 argY;
-
-    s16 var_s5 = 0;
-    s32 i = 0;
-    s16* yPosPtr = &PLAYER.posY.i.hi;
-    s16* xPosPtr = &PLAYER.posX.i.hi;
-    s32* vram_ptr = &g_Player.pl_vram_flag;
-
-    if (g_unkGraphicsStruct.unk18) {
-        *vram_ptr = 1;
-        return;
-    }
-    if ((PLAYER.velocityY == 0) && (g_Player.unk04 & 1) &&
-        !(g_Player.status &
-          (PLAYER_STATUS_BAT_FORM | PLAYER_STATUS_MIST_FORM))) {
-        var_s5 = 4;
-    } else {
-        i = 1;
-    }
-    for (; i < NUM_HORIZONTAL_SENSORS; i++, var_s5 = 0) {
-        if (g_Player.colFloor[i].effects & EFFECT_SOLID_FROM_BELOW) {
-            continue;
-        }
-        if (!((g_Player.colFloor[i].effects & EFFECT_UNK_0002) ||
-              (PLAYER.velocityY >= 0) ||
-              (g_Player.status & PLAYER_STATUS_TRANSFORM) ||
-              (abs(PLAYER.velocityX) >= FIX(2)))) {
-            continue;
-        }
-        var_s2 = g_Player.colFloor[i].effects &
-                 (EFFECT_UNK_8000 | EFFECT_UNK_0800 | EFFECT_SOLID);
-        if ((var_s2 == EFFECT_SOLID) || (var_s2 & EFFECT_UNK_0800)) {
-            if ((g_Player.status &
-                 (PLAYER_STATUS_BAT_FORM | PLAYER_STATUS_MIST_FORM)) &&
-                !(var_s2 & 1)) {
-                continue;
-            }
-            argX = *xPosPtr + g_SensorsFloor[i].x;
-            argY = *yPosPtr + g_SensorsFloor[i].y;
-            argY += (g_Player.colFloor[i].unk18 - 1);
-            CheckCollision(argX, argY, &sp10, 0);
-            if ((g_Player.status & PLAYER_STATUS_MIST_FORM) &&
-                (sp10.effects & EFFECT_MIST_ONLY)) {
-                sp10.effects &= ~(EFFECT_UNK_0002 | EFFECT_SOLID);
-            }
-            var_s1 = sp10.effects;
-            if (!(sp10.effects & EFFECT_SOLID)) {
-                if (((g_Player.colFloor[i].effects != EFFECT_SOLID) &&
-                     (g_Player.colFloor[i].effects !=
-                      (EFFECT_SOLID_FROM_ABOVE | EFFECT_SOLID))) ||
-                    (PLAYER.velocityY >= 0)) {
-                    if (var_s2 & EFFECT_UNK_0800) {
-                        *yPosPtr += var_s5 + g_Player.colFloor[i].unk8;
-                    } else {
-                        *yPosPtr += var_s5 + g_Player.colFloor[i].unk18;
-                    }
-                    *vram_ptr |= 1;
-                    return;
-                }
-                continue;
-            } else {
-                var_a0 =
-                    var_s1 & (EFFECT_UNK_8000 | EFFECT_UNK_0002 | EFFECT_SOLID);
-                if (var_a0 == (EFFECT_UNK_8000 | EFFECT_SOLID)) {
-                    if (i < 2) {
-                        *vram_ptr |=
-                            ((var_s1 & (EFFECT_UNK_4000 | EFFECT_UNK_2000 |
-                                        EFFECT_UNK_1000)) +
-                             0x7FFF + 2);
-                        *yPosPtr +=
-                            g_Player.colFloor[i].unk8 + sp10.unk18 - 1 + var_s5;
-                        return;
-                    }
-                    if ((i == 2) &&
-                        ((var_s1 &
-                          (EFFECT_UNK_8000 | EFFECT_UNK_4000 | EFFECT_SOLID)) ==
-                         (EFFECT_UNK_8000 | EFFECT_SOLID))) {
-                        g_Player.colFloor[2].effects = var_s1;
-                        g_Player.colFloor[2].unk10 = g_Player.colFloor[2].unk8;
-                    }
-                    if ((i == 3) &&
-                        ((var_s1 &
-                          (EFFECT_UNK_8000 | EFFECT_UNK_4000 | EFFECT_SOLID)) ==
-                         (EFFECT_UNK_8000 | EFFECT_UNK_4000 | EFFECT_SOLID))) {
-                        g_Player.colFloor[3].effects = var_s1;
-                        g_Player.colFloor[3].unk10 = g_Player.colFloor[3].unk8;
-                    }
-                }
-            }
-        }
-        if ((var_s2 != (EFFECT_UNK_8000 | EFFECT_SOLID)) || (i >= 2)) {
-            continue;
-        }
-        *vram_ptr |= g_Player.colFloor[i].effects &
-                     (EFFECT_UNK_8000 | EFFECT_UNK_4000 | EFFECT_UNK_2000 |
-                      EFFECT_UNK_1000 | EFFECT_SOLID);
-        *yPosPtr += var_s5 + g_Player.colFloor[i].unk18;
-        return;
-    }
-    if (g_Player.colFloor[1].effects & EFFECT_QUICKSAND) {
-        *vram_ptr |= 0x11;
-        if ((g_Timer & 3) == 0) {
-            (*yPosPtr)++;
-        }
-        return;
-    }
-    if (g_Player.colFloor[1].effects & EFFECT_WATER) {
-        *vram_ptr |= 0x80;
-    }
-    if (PLAYER.velocityY < 0) {
-        return;
-    }
-    argX = *xPosPtr + g_SensorsFloor[0].x;
-    argY = *yPosPtr + g_SensorsFloor[0].y + 10;
-    CheckCollision(argX, argY, &sp10, 0);
-    if ((sp10.effects & (EFFECT_UNK_8000 | EFFECT_SOLID)) != EFFECT_NONE) {
-        return;
-    }
-
-    for (i = 2; i < NUM_HORIZONTAL_SENSORS; i++) {
-        if ((g_Player.colFloor[3].effects & EFFECT_UNK_8000) &&
-            (g_Player.colFloor[2].effects & EFFECT_UNK_8000)) {
-            return;
-        }
-        var_s2 = g_Player.colFloor[i].effects;
-        sp30 =
-            ((var_s2 & (EFFECT_UNK_4000 | EFFECT_UNK_2000 | EFFECT_UNK_1000)) +
-             0x7FFF + 2);
-        if (!(var_s2 & EFFECT_UNK_8000)) {
-            continue;
-        }
-        if (i == 2) {
-            var_s1 = EFFECT_UNK_4000;
-            var_s6 = g_Player.colFloor[2].unk4;
-            var_s8 = var_s6 + 8;
-        } else {
-            var_s1 = 0;
-            var_s6 = g_Player.colFloor[3].unkC;
-            var_s8 = 8 - var_s6;
-        }
-        if ((var_s2 & EFFECT_UNK_4000) == var_s1) {
-            argX = var_s6 + (*xPosPtr + g_SensorsFloor[i].x);
-            argY = *yPosPtr + g_SensorsFloor[i].y;
-            CheckCollision(argX, argY, &sp10, 0);
-            if (sp10.effects & EFFECT_SOLID) {
-                *yPosPtr += sp10.unk18;
-                *vram_ptr |= sp30;
-                return;
-            }
-            continue;
-        }
-        if (var_s8 <= 0) {
-            continue;
-        }
-        if (!(var_s2 & 1)) {
-            continue;
-        }
-        argX = var_s6 + (*xPosPtr + g_SensorsFloor[i].x);
-        argY = *yPosPtr + g_SensorsFloor[i].y + g_Player.colFloor[i].unk10;
-        CheckCollision(argX, argY, &sp10, 0);
-        if (sp10.effects & EFFECT_SOLID) {
-            *yPosPtr += (sp10.unk18 + g_Player.colFloor[i].unk10);
-            *vram_ptr |= sp30;
-            return;
-        }
-    }
-}
-
-static void CheckCeiling(void) {
-    Collider collider;
-    s32 temp_fp;
-    u32 temp_s0;
-    s32 temp_v1;
-    s32 var_a0;
-    s16 var_a1;
-
-    s16 argX;
-    s16 argY;
-
-    u32 collidereffects;
-
-    s32 i = 1;
-
-    s16* yPosPtr = &PLAYER.posY.i.hi;
-    s16* xPosPtr = &PLAYER.posX.i.hi;
-    s32* vram_ptr = &g_Player.pl_vram_flag;
-
-    if (g_unkGraphicsStruct.unk18) {
-        return;
-    }
-    for (; i < NUM_HORIZONTAL_SENSORS; i++) {
-        if (g_Player.colCeiling[i].effects & EFFECT_SOLID_FROM_ABOVE) {
-            continue;
-        }
-        if ((g_Player.status & PLAYER_STATUS_MIST_FORM) &&
-            (collider.effects & EFFECT_MIST_ONLY)) {
-            collider.effects &= ~(EFFECT_UNK_0002 | EFFECT_SOLID);
-        }
-        temp_s0 = g_Player.colCeiling[i].effects &
-                  (EFFECT_UNK_8000 | EFFECT_UNK_0800 | EFFECT_SOLID);
-        if ((temp_s0 == EFFECT_SOLID) || (temp_s0 & EFFECT_UNK_8000)) {
-            if ((g_Player.status &
-                 (PLAYER_STATUS_BAT_FORM | PLAYER_STATUS_MIST_FORM)) &&
-                !(temp_s0 & EFFECT_SOLID)) {
-                continue;
-            }
-
-            argX = *xPosPtr + g_SensorsCeiling[i].x;
-            argY = *yPosPtr + g_SensorsCeiling[i].y +
-                   g_Player.colCeiling[i].unk10 + 1;
-            CheckCollision(argX, argY, &collider, 0);
-            // If in mist form, and collider is mist, then disable collision on
-            // that collider!
-            if (g_Player.status & PLAYER_STATUS_MIST_FORM &&
-                collider.effects & EFFECT_MIST_ONLY) {
-                collider.effects &= ~(EFFECT_UNK_0002 | EFFECT_SOLID);
-            }
-            collidereffects = collider.effects;
-            if (!(collidereffects & 1)) {
-                if ((g_Player.colCeiling[i].effects != EFFECT_SOLID) ||
-                    (PLAYER.velocityY <= 0)) {
-                    *vram_ptr |= 2;
-                    if (!(*vram_ptr & 1) &&
-                        ((g_Player.unk04 &
-                          (EFFECT_SOLID_FROM_ABOVE | EFFECT_SOLID)) !=
-                         (EFFECT_SOLID_FROM_ABOVE | EFFECT_SOLID))) {
-                        if (g_Player.colCeiling[i].effects & EFFECT_UNK_8000) {
-                            *yPosPtr += g_Player.colCeiling[i].unk10;
-                        } else {
-                            *yPosPtr += g_Player.colCeiling[i].unk20;
-                        }
-                    }
-                    return;
-                }
-                continue;
-            }
-            if ((collidereffects &
-                 (EFFECT_UNK_0800 | EFFECT_UNK_0002 | EFFECT_SOLID)) ==
-                (EFFECT_UNK_0800 | EFFECT_SOLID)) {
-                if (i < 2) {
-                    *vram_ptr |= (EFFECT_UNK_0800 | EFFECT_UNK_0002 |
-                                  ((collidereffects >> 4) &
-                                   (EFFECT_UNK_0400 | EFFECT_UNK_0200 |
-                                    EFFECT_UNK_0100)));
-                    if (!(*vram_ptr & 1)) {
-                        *yPosPtr +=
-                            g_Player.colCeiling[i].unk10 + collider.unk20 + 1;
-                    }
-                    return;
-                }
-                if ((i == 2) &&
-                    ((collidereffects &
-                      (EFFECT_UNK_4000 | EFFECT_UNK_0800 | EFFECT_SOLID)) ==
-                     (EFFECT_UNK_0800 | EFFECT_SOLID))) {
-                    g_Player.colFloor[2].effects = collidereffects;
-                    g_Player.colFloor[2].unk8 = g_Player.colFloor[2].unk10;
-                }
-                if ((i == 3) &&
-                    ((collidereffects &
-                      (EFFECT_UNK_4000 | EFFECT_UNK_0800 | EFFECT_SOLID)) ==
-                     (EFFECT_UNK_4000 | EFFECT_UNK_0800 | EFFECT_SOLID))) {
-                    g_Player.colFloor[3].effects = collidereffects;
-                    g_Player.colFloor[3].unk8 = g_Player.colFloor[3].unk10;
-                }
-            }
-            if ((collidereffects & EFFECT_UNK_0800) == EFFECT_NONE) {
-                *vram_ptr |=
-                    (EFFECT_UNK_0800 | EFFECT_UNK_0002 |
-                     ((collidereffects >> 4) &
-                      (EFFECT_UNK_0400 | EFFECT_UNK_0200 | EFFECT_UNK_0100)));
-                if (!(*vram_ptr & 1)) {
-                    *yPosPtr +=
-                        g_Player.colCeiling[i].unk10 + collider.unk20 + 1;
-                }
-                return;
-            }
-        }
-        if ((temp_s0 == (EFFECT_UNK_0800 | EFFECT_SOLID)) && (i < 2)) {
-            *vram_ptr |=
-                (EFFECT_UNK_0800 | EFFECT_UNK_0002 |
-                 ((g_Player.colCeiling[i].effects >> 4) &
-                  (EFFECT_UNK_0400 | EFFECT_UNK_0200 | EFFECT_UNK_0100)));
-            if (!(*vram_ptr & 1)) {
-                *yPosPtr += g_Player.colCeiling[i].unk20;
-            }
-            return;
-        }
-    }
-    if (PLAYER.velocityY > 0) {
-        return;
-    }
-    argX = *xPosPtr + g_SensorsCeiling[0].x;
-    argY = (*yPosPtr + g_SensorsCeiling[0].y) - 10;
-    CheckCollision(argX, argY, &collider, 0);
-    if ((collider.effects & EFFECT_SOLID) != EFFECT_NONE) {
-        return;
-    }
-    for (i = 2; i < NUM_HORIZONTAL_SENSORS; i++) {
-        if ((g_Player.colCeiling[3].effects & EFFECT_UNK_0800) &&
-            (g_Player.colCeiling[2].effects & EFFECT_UNK_0800)) {
-            return;
-        }
-        temp_s0 = g_Player.colCeiling[i].effects;
-        temp_fp = ((temp_s0 >> 4) &
-                   (EFFECT_UNK_0400 | EFFECT_UNK_0200 | EFFECT_UNK_0100)) +
-                  (EFFECT_UNK_0800 | EFFECT_UNK_0002);
-        if (temp_s0 & EFFECT_UNK_0800) {
-            if (i == 2) {
-                var_a0 = EFFECT_UNK_4000;
-                var_a1 = g_Player.colCeiling[2].unk4;
-                temp_v1 = var_a1 + 8;
-            } else {
-                var_a0 = 0;
-                var_a1 = g_Player.colCeiling[3].unkC;
-                temp_v1 = 8 - var_a1;
-            }
-            if ((temp_s0 & EFFECT_UNK_4000) == var_a0) {
-                argX = var_a1 + (*xPosPtr + g_SensorsCeiling[i].x);
-                argY = *yPosPtr + g_SensorsCeiling[i].y;
-                CheckCollision(argX, argY, &collider, 0);
-                if (collider.effects & EFFECT_SOLID) {
-                    *vram_ptr |= temp_fp;
-                    if (!(*vram_ptr & 1)) {
-                        *yPosPtr += collider.unk20;
-                    }
-                    return;
-                }
-            } else if ((temp_v1 > 0) && (temp_s0 & 1)) {
-                argX = var_a1 + (*xPosPtr + g_SensorsCeiling[i].x);
-                argY = *yPosPtr + g_SensorsCeiling[i].y +
-                       g_Player.colCeiling[i].unk8;
-                CheckCollision(argX, argY, &collider, 0);
-                if (collider.effects & EFFECT_SOLID) {
-                    if (!(*vram_ptr & 1)) {
-                        *yPosPtr +=
-                            collider.unk20 + g_Player.colCeiling[i].unk8;
-                    }
-                    *vram_ptr |= temp_fp;
-                    return;
-                }
-            }
-        }
-    }
-}
-
-static void CheckWallRight(void) {
-    Collider collider;
-    s32 temp_s0;
-    s32 i;
-
-    s16 argX;
-    s16 argY;
-
-    s16* yPosPtr = &PLAYER.posY.i.hi;
-    s16* xPosPtr = &PLAYER.posX.i.hi;
-    s32* vram_ptr = &g_Player.pl_vram_flag;
-
-    if (((g_StageId == STAGE_BO6) || (g_StageId == STAGE_RBO6) ||
-         (g_StageId == STAGE_DRE)) &&
-        (*xPosPtr >= 0xF8)) {
-        *xPosPtr = 0xF8;
-        *vram_ptr |= 4;
-        return;
-    }
-
-    if (g_unkGraphicsStruct.unk18) {
-        return;
-    }
-    temp_s0 =
-        g_Player.unk04 & (EFFECT_UNK_8000 | EFFECT_UNK_4000 | EFFECT_UNK_0800 |
-                          EFFECT_UNK_0400 | EFFECT_UNK_0002 | EFFECT_SOLID);
-    if ((temp_s0 == (EFFECT_UNK_8000 | EFFECT_UNK_0002 | EFFECT_SOLID)) ||
-        (temp_s0 == (EFFECT_UNK_0800 | EFFECT_UNK_0002 | EFFECT_SOLID)) ||
-        (temp_s0 == (EFFECT_UNK_8000 | EFFECT_UNK_0800 | EFFECT_UNK_0002 |
-                     EFFECT_SOLID))) {
-        *vram_ptr |= 4;
-        return;
-    }
-
-    for (i = 0; i < NUM_VERTICAL_SENSORS; i++) {
-        temp_s0 = g_Player.colWall[i].effects &
-                  (EFFECT_UNK_8000 | EFFECT_UNK_4000 | EFFECT_UNK_0800 |
-                   EFFECT_UNK_0002 | EFFECT_SOLID);
-        if ((temp_s0 == (EFFECT_UNK_8000 | EFFECT_UNK_4000 | EFFECT_SOLID)) ||
-            (temp_s0 == (EFFECT_UNK_8000 | EFFECT_UNK_4000 | EFFECT_UNK_0002 |
-                         EFFECT_SOLID)) ||
-            (temp_s0 == (EFFECT_UNK_4000 | EFFECT_UNK_0800 | EFFECT_SOLID)) ||
-            (temp_s0 == (EFFECT_UNK_4000 | EFFECT_UNK_0800 | EFFECT_UNK_0002 |
-                         EFFECT_SOLID)) ||
-            (temp_s0 == (EFFECT_UNK_8000 | EFFECT_UNK_0002 | EFFECT_SOLID)) ||
-            (temp_s0 == (EFFECT_UNK_0800 | EFFECT_UNK_0002 | EFFECT_SOLID)) ||
-            (temp_s0 == (EFFECT_UNK_0002 | EFFECT_SOLID))) {
-            argX = *xPosPtr + g_SensorsWall[i].x + g_Player.colWall[i].unk4 - 1;
-            argY = *yPosPtr + g_SensorsWall[i].y;
-            CheckCollision(argX, argY, &collider, 0);
-            if ((collider.effects & EFFECT_SOLID) == EFFECT_NONE) {
-                *vram_ptr |= 4;
-                *xPosPtr += g_Player.colWall[i].unk4;
-                return;
-            }
-        }
-
-        if (!(*vram_ptr & 1)) {
-            if ((temp_s0 & (EFFECT_UNK_8000 | EFFECT_UNK_4000 |
-                            EFFECT_UNK_0800)) == EFFECT_UNK_8000 &&
-                (i != 0) &&
-                ((g_Player.colWall[0].effects & EFFECT_UNK_0800) ||
-                 !(g_Player.colWall[0].effects &
-                   (EFFECT_UNK_8000 | EFFECT_UNK_0800 | EFFECT_UNK_0002)))) {
-                *vram_ptr |= 4;
-                *xPosPtr += g_Player.colWall[i].unk4;
-                return;
-            }
-            if (((temp_s0 & (EFFECT_UNK_8000 | EFFECT_UNK_4000 |
-                             EFFECT_UNK_0800)) == EFFECT_UNK_0800) &&
-                (i != 6) &&
-                ((g_Player.colWall[6].effects & EFFECT_UNK_8000) ||
-                 !(g_Player.colWall[6].effects &
-                   (EFFECT_UNK_8000 | EFFECT_UNK_0800 | EFFECT_UNK_0002)))) {
-                *vram_ptr |= 4;
-                *xPosPtr += g_Player.colWall[i].unk4;
-                return;
-            }
-        }
-    }
-}
-
-static void CheckWallLeft(void) {
-    Collider collider;
-    s32 temp_s0;
-    s32 i;
-    s16 argX;
-    s16 argY;
-
-    s16* yPosPtr = &PLAYER.posY.i.hi;
-    s16* xPosPtr = &PLAYER.posX.i.hi;
-    s32* vram_ptr = &g_Player.pl_vram_flag;
-
-    if (((g_StageId == STAGE_BO6) || (g_StageId == STAGE_RBO6) ||
-         (g_StageId == STAGE_DRE)) &&
-        (*xPosPtr <= 8)) {
-        *xPosPtr = 8;
-        *vram_ptr |= 8;
-        return;
-    }
-
-    if (g_unkGraphicsStruct.unk18) {
-        return;
-    }
-    temp_s0 =
-        g_Player.unk04 & (EFFECT_UNK_8000 | EFFECT_UNK_4000 | EFFECT_UNK_0800 |
-                          EFFECT_UNK_0400 | EFFECT_UNK_0002 | EFFECT_SOLID);
-    if ((temp_s0 == (EFFECT_UNK_8000 | EFFECT_UNK_4000 | EFFECT_UNK_0002 |
-                     EFFECT_SOLID)) ||
-        (temp_s0 == (EFFECT_UNK_0800 | EFFECT_UNK_0400 | EFFECT_UNK_0002 |
-                     EFFECT_SOLID)) ||
-        (temp_s0 == (EFFECT_UNK_8000 | EFFECT_UNK_4000 | EFFECT_UNK_0800 |
-                     EFFECT_UNK_0400 | EFFECT_UNK_0002 | EFFECT_SOLID))) {
-        *vram_ptr |= 8;
-        return;
-    }
-    for (i = NUM_VERTICAL_SENSORS; i < NUM_VERTICAL_SENSORS * 2; i++) {
-        temp_s0 = g_Player.colWall[i].effects &
-                  (EFFECT_UNK_8000 | EFFECT_UNK_4000 | EFFECT_UNK_0800 |
-                   EFFECT_UNK_0002 | EFFECT_SOLID);
-        if ((temp_s0 == (EFFECT_UNK_8000 | EFFECT_SOLID)) ||
-            (temp_s0 == (EFFECT_UNK_8000 | EFFECT_UNK_0002 | EFFECT_SOLID)) ||
-            (temp_s0 == (EFFECT_UNK_0800 | EFFECT_SOLID)) ||
-            (temp_s0 == (EFFECT_UNK_0800 | EFFECT_UNK_0002 | EFFECT_SOLID)) ||
-            (temp_s0 == (EFFECT_UNK_8000 | EFFECT_UNK_4000 | EFFECT_UNK_0002 |
-                         EFFECT_SOLID)) ||
-            (temp_s0 == (EFFECT_UNK_4000 | EFFECT_UNK_0800 | EFFECT_UNK_0002 |
-                         EFFECT_SOLID)) ||
-            (temp_s0 == (EFFECT_UNK_0002 | EFFECT_SOLID))) {
-            argX = *xPosPtr + g_SensorsWall[i].x + g_Player.colWall[i].unkC + 1;
-            argY = *yPosPtr + g_SensorsWall[i].y;
-            CheckCollision(argX, argY, &collider, 0);
-            if ((collider.effects & EFFECT_SOLID) == EFFECT_NONE) {
-                *vram_ptr |= 8;
-                *xPosPtr += g_Player.colWall[i].unkC;
-                return;
-            }
-        }
-        if (!(*vram_ptr & 1)) {
-            if (((temp_s0 &
-                  (EFFECT_UNK_8000 | EFFECT_UNK_4000 | EFFECT_UNK_0800)) ==
-                 (EFFECT_UNK_8000 | EFFECT_UNK_4000)) &&
-                (i != 7) &&
-                ((g_Player.colWall[NUM_VERTICAL_SENSORS].effects &
-                  EFFECT_UNK_0800) ||
-                 !(g_Player.colWall[NUM_VERTICAL_SENSORS].effects &
-                   (EFFECT_UNK_8000 | EFFECT_UNK_0800 | EFFECT_UNK_0002)))) {
-                *vram_ptr |= 8;
-                *xPosPtr += g_Player.colWall[i].unkC;
-                return;
-            }
-            if (((temp_s0 &
-                  (EFFECT_UNK_8000 | EFFECT_UNK_4000 | EFFECT_UNK_0800)) ==
-                 (EFFECT_UNK_4000 | EFFECT_UNK_0800)) &&
-                (i != 13) &&
-                ((g_Player.colWall[NUM_VERTICAL_SENSORS + 6].effects &
-                  EFFECT_UNK_8000) ||
-                 !(g_Player.colWall[NUM_VERTICAL_SENSORS + 6].effects &
-                   (EFFECT_UNK_8000 | EFFECT_UNK_0800 | EFFECT_UNK_0002)))) {
-                *vram_ptr |= 8;
-                *xPosPtr += g_Player.colWall[i].unkC;
-                return;
-            }
-        }
     }
 }
